refactor(test): move parse-capture helpers into the shared harness
read_text(), record(), StderrToFile, parse_records(), and logged() were local to
test_variant.cpp but are not specific to it, so they move to test_helpers, taking
ParseResult with them since parse_records() and logged() both name it.

StderrToFile's bodies move out of line to keep the header @brief-only. The record
loop variable in write_tmp_vcf() is renamed to rec, so the new record() is not
shadowed in that translation unit. The chr1-specific assertions kept_on_hap(),
total_kept(), and wrote_pos() stay in test_variant.cpp.
